requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
confection==0.1.3
contourpy==1.1.1
cycler==0.11.0
cymem==2.0.7
distlib==0.3.7
elastic-transport==8.4.0
elasticsearch==8.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.12.3
flake8==6.1.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.17.1
identify==2.5.27
idna==3.4
#importlib-metadata==6.8.0
#metadata>=0.1,<0.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.3.0
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy==1.5.1
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
#numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.0
packaging==23.1
#pandas==2.1.0
pathspec==0.11.2
pathy==0.10.2
Pillow==9.5.0
platformdirs==3.10.0
pre-commit==3.4.0
preshed==3.0.8
protobuf==4.24.3
pyarrow==13.0.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydeck==0.8.0
pyflakes==3.1.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PySide6==6.4.0
PySide6-Addons==6.4.0
PySide6-Essentials==6.4.0
pysrt==1.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.0
#scipy==1.11.2
shiboken6==6.4.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
pytesseract
langchain
pyvis
spacy_streamlit
SentencePiece
tiktoken
pdf2image
python-docx